[feature] 增加当前缴费人员功能，完善费用检查条件, 完善email群体发送
